Tweak popover design
